pyproject.toml
[project]
name = "aidd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "langchain-huggingface>=0.1.2",
    "langgraph>=0.2.68",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "polars>=1.19.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "streamlit>=1.52.2",
]
